fixed #4231: Added "DisableLogging" registry setting

+++ ShareX/CleanupManager.cs
@@ -74,16 +74,19 @@ namespace ShareX
 
         private static void CleanupFolder(string folderPath, string fileNamePattern, int keepFileCount)
         {
-            DirectoryInfo directoryInfo = new DirectoryInfo(folderPath);
-
-            IEnumerable<FileInfo> files = directoryInfo.GetFiles(fileNamePattern).
-                OrderByDescending(f => f.LastWriteTime.Year <= 1601 ? f.CreationTime : f.LastWriteTime).Skip(keepFileCount);
-
-            foreach (FileInfo file in files)
+            if (Directory.Exists(folderPath))
             {
-                file.Delete();
+                DirectoryInfo directoryInfo = new DirectoryInfo(folderPath);
 
-                DebugHelper.WriteLine($"File deleted: {file.FullName}");
+                IEnumerable<FileInfo> files = directoryInfo.GetFiles(fileNamePattern).
+                    OrderByDescending(f => f.LastWriteTime.Year <= 1601 ? f.CreationTime : f.LastWriteTime).Skip(keepFileCount);
+
+                foreach (FileInfo file in files)
+                {
+                    file.Delete();
+
+                    DebugHelper.WriteLine($"File deleted: {file.FullName}");
+                }
             }
         }
